+++ src/utils/common.js
@@ -1,6 +1,3 @@
-import dayjs from 'dayjs';
-import { DATE_FORMAT, SHORT_DATE_FORMAT, TIME_FORMAT } from './const.js';
-
 const getRandomArrayElement = (items) => items[Math.floor(Math.random() * items.length)];
 
 const getRandomInteger = (max) => Math.floor(Math.random() * max);
@@ -14,10 +11,6 @@ const createIdGenerator = () => {
   };
 };
 
-const humanizePointsDate = (date) => dayjs(date).format(DATE_FORMAT);
-const humanizeShortDate = (date) => dayjs(date).format(SHORT_DATE_FORMAT).toUpperCase();
-const humanizeTime = (date) => dayjs(date).format(TIME_FORMAT);
-
 const capitalizeFirstLetter = (string) => string.charAt(0).toUpperCase() + string.slice(1);
 
 const isEscapeKey = (evt) => evt.key === 'Escape';
@@ -25,8 +18,5 @@ const isEscapeKey = (evt) => evt.key === 'Escape';
 export { getRandomArrayElement,
   getRandomInteger,
   createIdGenerator,
-  humanizeTime,
-  humanizePointsDate,
-  humanizeShortDate,
   capitalizeFirstLetter,
   isEscapeKey };
