+++ src/utils/dates.js
@@ -1,0 +1,12 @@
+import dayjs from 'dayjs';
+import { DATE_FORMAT, SHORT_DATE_FORMAT, TIME_FORMAT } from '../const.js';
+
+const humanizePointsDate = (date) => dayjs(date).format(DATE_FORMAT);
+const humanizeShortDate = (date) => dayjs(date).format(SHORT_DATE_FORMAT).toUpperCase();
+const humanizeTime = (date) => dayjs(date).format(TIME_FORMAT);
+
+export {
+  humanizeTime,
+  humanizePointsDate,
+  humanizeShortDate
+};
