+++ src/view/no-point-view.js
@@ -1,0 +1,9 @@
+import AbstractView from '../framework/view/abstract-view.js';
+
+const createNoViewTemplate = () => '<p class="trip-events__msg">Click New Event to create your first point</p>';
+
+export default class NoPointView extends AbstractView {
+  get template() {
+    return createNoViewTemplate();
+  }
+}
